+++ src/utils/storage/types/index.ts
@@ -1,4 +1,3 @@
-// Chrome Storage 관련 타입 정의
 export interface ExtensionSettings {
   autoGenerate: boolean;
   reportFormat: "markdown" | "pdf" | "html";
